Create an API and add the following functionalities : - Add a task - Get all tasks

+++ src/main/java/com/computerience/todoapp/repository/TaskRepository.java
@@ -1,7 +1,8 @@
 package com.computerience.todoapp.repository;
 
+
 import com.computerience.todoapp.entity.Task;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TaskRepository extends JpaRepository<Task, Integer> {
+public interface TaskRepository extends JpaRepository<Task, Long> {
 }
